refactor: extract storefront, theme, and product info concerns
- Extract storefront logic to src/storefront.ts (buildAppleMusicURL,
  extractStorefrontFromURL, handleStorefrontNavigation)
- Extract theme management to src/theme.ts (initCatppuccinCSS, state)
- Add getProductInfo() to src/paths.ts to encapsulate package.json reads
- Remove pkg require from tray.ts, replace with getProductInfo()
- Reduce setupContentHandlers parameter list from 6 to 4 by passing
  assets object
- Update test import for extracted URL functions
- Fixes inappropriate intimacy, divergent change, and long parameter
  list smells
