+++ src/paths.ts
@@ -7,3 +7,32 @@ export function getAssetPath(...parts: string[]): string {
     : path.join(__dirname, '..');
   return path.join(base, ...parts);
 }
+
+export interface ProductInfo {
+  productName: string;
+  description: string;
+  author: string;
+  license: string;
+}
+
+let cachedProductInfo: ProductInfo | null = null;
+
+export function getProductInfo(): ProductInfo {
+  if (cachedProductInfo) {
+    return cachedProductInfo;
+  }
+
+  const pkg = require(path.join(__dirname, '..', 'package.json'));
+  const author = typeof pkg.author === 'string'
+    ? pkg.author.replace(/\s*<[^>]+>/, '')
+    : (pkg.author?.name ?? '');
+
+  cachedProductInfo = {
+    productName: app.getName(),
+    description: pkg.description ?? '',
+    author,
+    license: pkg.license ?? '',
+  };
+
+  return cachedProductInfo;
+}
